refactor(profile): extract appbar to local widget
